+++ src/commands/restart.py
@@ -18,6 +18,7 @@ from src.utils.doco import ProjectInfo
 class Options(DownOptions):
     do_pull: bool
     do_log: bool
+    no_build: bool
     dry_run: bool
 
 
@@ -40,7 +41,7 @@ def restart_project(project: ComposeProject, options: Options, info: ProjectInfo
         project.file,
         command=[
             "up",
-            "--build",
+            *(["--build"] if not options.no_build else []),
             *(["--pull", "always"] if options.do_pull else []),
             "-d",
         ],
@@ -65,10 +66,11 @@ def main(  # noqa: CFQ002 (max arguments)
     remove_volumes: bool = typer.Option(
         False, "--remove-volumes", "-v", help="Remove volumes (implies -f / --force)."
     ),
-    no_remove_orphans: bool = typer.Option(False, "--no-remove-orphans", "-k", help="Keep orphans."),
+    no_remove_orphans: bool = typer.Option(False, "--no-remove-orphans", help="Keep orphans."),
     force: bool = typer.Option(False, "--force", "-f", help="Force calling down even if not running."),
     do_pull: bool = typer.Option(False, "--pull", "-p", help="Pull images before running."),
     do_log: bool = typer.Option(False, "--log", "-l", help="Also show logs."),
+    no_build: bool = typer.Option(False, "--no-build", help="Don't build images before running."),
     dry_run: bool = typer.Option(
         False, "--dry-run", "-n", help="Do not actually stop anything, only show what would be done."
     ),
@@ -96,6 +98,7 @@ def main(  # noqa: CFQ002 (max arguments)
                     force_down=force,
                     do_pull=do_pull,
                     do_log=do_log,
+                    no_build=no_build,
                     dry_run=dry_run,
                 ),
                 info=info,